Update memes
I finally got around to organizing the (used-to-be) full list. I made a Python program that automated as much of it as possible, because man it is tedious. I just finished scraping Wombo again for all the new memes (and Wombo Combos) that have been added since I first scraped it in April, so those will be coming relatively soon once I'm through with sorting them as well.